v1.5 — Logo Generator + Vector Generator modules
- Logo Generator: AI SVG logos with style picker, live preview, favicon auto-generation
- Vector Generator: AI SVG icons, illustrations, badges, patterns
- SVG rendering via resvg/usvg, ICO generation via ico crate
- Export: SVG, multi-size PNG, ICO, favicon package, webmanifest, ZIP
- Both modules under Team license tier
- Sidebar: 🎨 Logo Generator + 📐 Vector Generator (Library section)
- cargo check: 0 errors, cargo clippy: 0 warnings

+++ src/vector_renderer.rs
@@ -23,7 +23,7 @@ pub fn render_to_image(svg_data: &str, scale: f32) -> Result<DynamicImage, Strin
 
     let render_tree = Tree::from_usvg(&tree);
     render_tree.render(
-        resvg::tiny_skia::Transform::from_scale(scale as f32, scale as f32),
+        resvg::tiny_skia::Transform::from_scale(scale, scale),
         &mut pixmap.as_mut(),
     );
